requirements.txt
opencv-python==4.5.3.56
pillow==8.3.1
imageio==2.9.0
numpy==1.20.3
tqdm==4.61.2
tensorboardX==2.4
torchsummary==1.5.1